ARROW-16590: Consolidate utilities involving row-major format.  Add tests for encoding and decoding rows as well as a test for comparing column-major data with row-major data.
