+++ cpp/src/arrow/compute/light_array.h
@@ -21,8 +21,10 @@
 
 #include "arrow/array.h"
 #include "arrow/compute/exec.h"
+#include "arrow/compute/exec/util.h"
 #include "arrow/type.h"
 #include "arrow/util/logging.h"
+#include "arrow/util/cpu_info.h"
 
 /// This file contains lightweight containers for Arrow buffers.  These containers
 /// makes compromises in terms of strong ownership and the range of data types supported
@@ -30,6 +32,12 @@
 
 namespace arrow {
 namespace compute {
+
+struct LightContext {
+  bool has_avx2() const { return (hardware_flags & arrow::internal::CpuInfo::AVX2) > 0; }
+  int64_t hardware_flags;
+  util::TempVectorStack* stack;
+};
 
 /// \brief Description of the layout of a "key" column
 ///
